First attempt to fix issue #5
Issue #5: decrement time to weird numbers
Can't decrease a digit below 0... sort of. When I have, for example 00:10 and I decrement the minutes. It should go down to 00:09, but intead it stays at 00:10. Fortunately, I know exactly what's causing this: the way that I have minutes, tens of minutes, hours, tens of hours.